requirements.txt
google-cloud-storage>=2.10.0
google-cloud-speech>=2.24.0
functions-framework>=3.5.0
google-generativeai>=0.5.0
